+++ app/skel/index.php
@@ -1,0 +1,25 @@
+<?php
+
+/**
+ * Iggy Rapid Prototyping App
+ *
+ * @license http://opensource.org/licenses/MIT
+ * @link https://github.com/caseyamcl/iggy
+ * @version 1.0
+ * @package caseyamcl/iggy
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ *
+ * ------------------------------------------------------------------
+ */
+
+/**
+ * Iggy Main Index File
+ *
+ * @author Casey McLaughlin <caseyamcl@gmail.com>
+ */
+
+// Just one line; require the bootstrap in the Application folder
+require_once __DIR__ . '/app/bootstrap.php';