+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 PHPOAIPMH Changelog
 ===================
 
+## Version 0.6 (2015 Feb 11)
+- Added `.gitattributes`
+- Changed to PSR-4 Autloader Standard
+- Added `InstallIggy` and Composer `create-project` hook
+- Changed documentation for installation
+
 ## Version 0.5.1 (2015 Feb 11)
 - Improved documentation